feat(frontend): add loader to block of routes ss-270
