+++ apps/frontend/src/libs/components/explore-routes-block/libs/helpers/get-current-user-position.helper.ts
@@ -1,0 +1,15 @@
+const getCurrentUserPosition = (): Promise<[number, number]> => {
+	return new Promise((resolve, reject) => {
+		// eslint-disable-next-line sonarjs/no-intrusive-permissions
+		navigator.geolocation.getCurrentPosition(
+			(position) => {
+				resolve([position.coords.latitude, position.coords.longitude]);
+			},
+			(error) => {
+				reject(new Error(`Failed to get position: ${error.message}`));
+			},
+		);
+	});
+};
+
+export { getCurrentUserPosition };
