+++ apps/frontend/src/libs/components/explore-routes-block/libs/helpers/helpers.ts
@@ -1,3 +1,4 @@
+export { getCurrentUserPosition } from "./get-current-user-position.helper.js";
 export { getDistance } from "./get-distance.helper.js";
 export { getPOICoordinates } from "./get-poi-coordinates.helper.js";
 export { getRoutesForPOIs } from "./get-routes-for-pois.helper.js";
